renamed main detection eval file
